BREAKING CHANGE! Gulp and browserify
* The move to gulp has not created any big change.
* The move to use browserify has caused a mayor rewrite. The use of
globals are frowned upon so the global game and player is now gone. game
should be supplied to objects that need it. This has been done.
* All objects in folder agent and objects has been updated to take a
game argument. This means that all uses of the objects have to now have
the game argument in the first place.
* player is injected into the game object. Use game.player.
* The file structure has been updated because browserify's
require-system does not rely on any structural setup.
* A SuperState has been added which takes care of checking for audio
decoding, making it more robust than before. It also takes care of
shutdown.